+++ sample-homecharge/src/main/java/com/example/homecharge/ChargeActivity.java
@@ -29,7 +29,6 @@ import com.squareup.sdk.pos.TransactionRequest;
 import com.squareup.sdk.pos.PosClient;
 import com.squareup.sdk.pos.PosSdk;
 
-import static com.squareup.sdk.pos.CurrencyCode.USD;
 import static com.squareup.sdk.pos.TransactionRequest.TenderType.CARD_FROM_READER;
 
 public class ChargeActivity extends AppCompatActivity {
@@ -45,7 +44,7 @@ public class ChargeActivity extends AppCompatActivity {
   }
 
   public void startTransaction(int dollarAmount, String note) {
-    TransactionRequest request = new TransactionRequest.Builder(dollarAmount * 1_00, USD).note(note)
+    TransactionRequest request = new TransactionRequest.Builder(dollarAmount * 1_00).note(note)
         .autoReturn(true)
         .restrictTendersTo(CARD_FROM_READER)
         .build();