+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Change Log
 * Replaced the CARD tender type with CARD_FROM_READER, KEYED_IN_CARD, and GIFT_CARD.
 * Added an optional "skipReceipt" which allows configuring whether or not to skip the receipt screen
   on a per-transaction basis.
+* Removed the currency code parameter. The transaction will be processed with the currency of the
+  logged-in merchant account in Point of Sale.
 
 ## Version 2.0